refactor rotary to use layers vs transformers pattern + add warning to readme
